Refactored into lib crate for core functions
